change damage system
* algorithm of taken damage is different for different target's type
* change EnemyType to Enum

+++ src/main/java/automatic/fire/complex/units/enemy/Infantry.java
@@ -2,7 +2,7 @@ package automatic.fire.complex.units.enemy;
 
 public class Infantry extends Enemy {
 
-    public Infantry(int posX, int posY, double protectionLevel) {
+    public Infantry(int posX, int posY, int protectionLevel) {
         super(posX, posY, protectionLevel);
     }
 }
