+++ src/main/java/automatic/fire/complex/units/enemy/Tank.java
@@ -2,7 +2,7 @@ package automatic.fire.complex.units.enemy;
 
 public class Tank extends Enemy {
 
-    public Tank(int posX, int posY, double protectionLevel) {
+    public Tank(int posX, int posY, int protectionLevel) {
         super(posX, posY, protectionLevel);
     }
